refactor(test): 공통 테스트 클래스 package 이동
